+++ Sources/KeyboardKit/Layout/KeyboardLayoutConfiguration.swift
@@ -62,7 +62,7 @@ public extension KeyboardLayoutConfiguration {
         standard(
             forDevice: context.deviceType,
             screenSize: context.screenSize,
-            orientation: context.screenOrientation)
+            orientation: context.interfaceOrientation)
         #else
         .standardPhoneLandscape
         #endif
@@ -75,7 +75,7 @@ public extension KeyboardLayoutConfiguration {
     static func standard(
         forDevice device: DeviceType,
         screenSize size: CGSize,
-        orientation: UIInterfaceOrientation
+        orientation: InterfaceOrientation
     ) -> KeyboardLayoutConfiguration {
         switch device {
         case .pad: return standardPad(forScreenSize: size, orientation: orientation)
@@ -88,7 +88,7 @@ public extension KeyboardLayoutConfiguration {
      */
     static func standardPad(
         forScreenSize size: CGSize,
-        orientation: UIInterfaceOrientation
+        orientation: InterfaceOrientation
     ) -> KeyboardLayoutConfiguration {
         let isPortrait = orientation.isPortrait
         if size.isScreenSize(.iPadProLargeScreenPortrait) {
@@ -104,7 +104,7 @@ public extension KeyboardLayoutConfiguration {
      */
     static func standardPhone(
         forScreenSize size: CGSize,
-        orientation: UIInterfaceOrientation
+        orientation: InterfaceOrientation
     ) -> KeyboardLayoutConfiguration {
         let isPortrait = orientation.isPortrait
         if size.isScreenSize(.iPhoneProMaxScreenPortrait) {