+++ Sources/KeyboardKit/Layout/Providers/iPhoneKeyboardLayoutProvider.swift
@@ -193,11 +193,7 @@ private extension iPhoneKeyboardLayoutProvider {
     }
 
     func isPortrait(_ context: KeyboardContext) -> Bool {
-        #if os(iOS)
-        return context.screenOrientation.isPortrait
-        #else
-        return false
-        #endif
+        context.interfaceOrientation.isPortrait
     }
 
     /**