+++ Sources/KeyboardKit/_Deprecated/Layout+Deprecated.swift
@@ -156,11 +156,11 @@ public extension KeyboardLayoutConfiguration {
     /**
      The standard config for the provided idiom and screen.
      */
-    @available(*, deprecated, message: "Use standard(forDevice:screenSize:oriantation) instead")
+    @available(*, deprecated, message: "Use standard(forDevice:screenSize:orientation) instead")
     static func standard(
         forIdiom idiom: UIUserInterfaceIdiom,
         screenSize size: CGSize,
-        orientation: UIInterfaceOrientation) -> KeyboardLayoutConfiguration {
+        orientation: InterfaceOrientation) -> KeyboardLayoutConfiguration {
         switch idiom {
         case .pad: return standardPad(forScreenSize: size, orientation: orientation)
         default: return standardPhone(forScreenSize: size, orientation: orientation)